Moving more buffers into Coreneuron class
- NgetBufReceive buffer
- bamech
- nrn_prop_(d)param_size_
- buffers for artificial cell

Note: For now the accesses to any of those vectors is through a reference
getter, which is really not how it should be done, but in the next step
many of the involved functions will become methods that access the
private data. In other cases the data will be rearranged and accessed
through it's own container class.